# src/main/resources/swagger.yaml
...
openapi: 3.0.1
info:
  title: E-Commerce API 문서
  description: MockAPI 데이터 연결 테스트
  version: v1.0.0
servers:
  - url: http://localhost:8080
    description: Generated server url
...
paths:
...
  "/api/coupons/use":
    post:
      tags:
        - 쿠폰
      summary: 사용자 쿠폰 사용
      description: 사용자 쿠폰 사용 하는 API
      operationId: useCoupon
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/UseCouponRequest"
        required: true
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/UseCouponResponse"
  "/api/coupons/issue":
    post:
      tags:
        - 쿠폰
      summary: 쿠폰 발급
      description: 쿠폰 발급 하는 API
      operationId: issueCoupon
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/IssueCouponRequest"
        required: true
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/IssueCouponResponse"
...
  "/api/coupons/owned/{memberId}":
    get:
      tags:
        - 쿠폰
      summary: 사용자 쿠폰 목록 조회
      description: 사용자 쿠폰 목록 조회 하는 API
      operationId: getOwnedCoupons
      parameters:
        - name: memberId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/OwnedCouponResponse"
...
components:
  schemas:
...
    UseCouponRequest:
      type: object
      properties:
        memberId:
          type: integer
          format: int64
        couponId:
          type: integer
          format: int64
    UseCouponResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
    IssueCouponRequest:
      type: object
      properties:
        memberId:
          type: integer
          format: int64
        couponId:
          type: integer
          format: int64
    IssueCouponResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
...
    OwnedCouponResponse:
      type: object
      properties:
        couponId:
          type: integer
          format: int64
        couponName:
          type: string
        expired:
          type: boolean
...